chore: detach files and remove unnecessary files
