--- tig/lib/presentation/widgets/buttons/menu_button.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MenuButton extends StatelessWidget {
-  final String text;
-  final VoidCallback onPressed;
-
-  const MenuButton({super.key, required this.text, required this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: onPressed,
-      child: Text(text),
-    );
-  }
-}
