+++ tig/lib/presentation/widgets/styles/fixed_scroll_behavior.dart
@@ -1,9 +1,5 @@
 import 'package:flutter/material.dart';
 
-class Helpers {
-  static FixedScrollBehavior get fixedScrollBehavior => FixedScrollBehavior();
-}
-
 class FixedScrollBehavior extends ScrollBehavior {
   @override
   ScrollPhysics getScrollPhysics(BuildContext context) =>
